requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
dominate==2.9.1
efficientnet_pytorch==0.7.1
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
lazy_loader==0.4
lpips==0.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pretrainedmodels==0.7.4
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.13.1
segmentation_models_pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@5a1a8f1088a96a71cc7ffc025d14f41281ea1263
six==1.16.0
sympy==1.13.1
tifffile==2024.9.20
timm==1.0.11
torch==2.5.1
torcheval==0.0.7
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.6
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
visdom==0.2.4
websocket-client==1.8.0
